기존의 chatting room components로 나누기
