+++ frontend/src/Components/ChattingRoom/ChatLeft.jsx
@@ -1,0 +1,15 @@
+import React, { useState, useRef, useEffect } from "react";
+import style from "./ChattingRoom.module.css";
+
+
+const ChatLeft = () => {
+
+
+    return (
+        <div className={style.Room_left}>
+            
+        </div>
+    );
+};
+
+export default ChatLeft;
