+++ frontend/src/Components/ChattingRoom/ChatRightTop.jsx
@@ -1,0 +1,19 @@
+import React, { useState, useRef, useEffect } from "react";
+import style from "./ChattingRoom.module.css";
+
+
+const ChatRightTop = () => {
+
+
+    return (
+        <div className={style.Room_right_top}>
+            {/*  
+            1. 자물쇠 비밀방
+            2. 제목
+            3. 게임 정보?
+            */}
+        </div>
+    );
+};
+
+export default ChatRightTop;
