+++ frontend/src/Components/ChattingRoom/ChattingRoom.module.css
@@ -1,0 +1,120 @@
+/* chatRoom */
+
+.Room_vertical {
+    display: flex;
+    justify-content: center;
+    height: 90vh;
+}
+
+.Room_container {
+    display: flex;
+    flex: 1;
+}
+
+.Room_left{
+    border: 1px solid #333;
+    flex: 1;
+    margin: 0 10px;
+    border-radius: 30px;
+}
+
+.Room_right {
+    flex: 2;
+    margin: 0 10px;
+    display: flex;
+    flex-direction: column;
+}
+
+
+.Room_right_top{
+    border: 1px solid #333;
+    flex: 1;
+    margin-bottom: 20px ;
+    border-radius: 30px;
+}
+
+
+
+
+
+
+
+
+.Room_right_bottom {
+    border: 1px solid #333;
+    height: 70%;
+    border-radius: 30px;
+    display: flex;
+    justify-content: center;
+    align-items: flex-end;
+    position: relative; 
+    z-index: 0; 
+}
+
+.Room_right_bottom::before {
+    content: "";
+    position: absolute;
+    top: 0;
+    left: 0;
+    right: 0;
+    bottom: 0;
+    background-image: url('../../Assets/images/logo.png');
+    background-position: center;
+    background-size: 300px;
+    background-repeat: no-repeat;
+    opacity: 0.3; 
+    z-index: -1; 
+    /* border-radius: 30px;  */
+}
+
+
+
+.Room_bottom_submit {
+    width: 80%;
+    margin-bottom: 10px;
+    border: 2px solid;
+    border-radius: 30px;
+}
+
+.Room_bottom_submit_form {
+    display: flex;
+    align-items: flex-end;
+}
+
+
+
+.Room_bottom_submit_textarea {
+    flex: 1;
+    padding: 10px;
+    margin: 5px 10px;
+    border: none;
+    font-size: 16px;
+    overflow-y: auto; 
+    min-height: 20px; 
+    max-height: 200px; 
+    resize: none;
+  }
+
+.Room_bottom_submit_textarea:focus {
+    outline: none;
+    color: #000;
+}
+
+.Room_bottom_submit_button {
+    width: 40px;
+    height: 40px;
+    color: black;
+    border: 0;
+    border-radius: 30px;
+    background-color: #ccc;
+    transform: rotate(-90deg);
+    padding-top: 3px;
+    margin: 5px 10px;
+    font-size: 20px;
+    color: #fff;
+}
+
+.Room_bottom_submit_button:hover{
+    cursor: pointer;
+}
+
